+++ integration_tests/webserver/test_rest_connector.py
@@ -12,12 +12,10 @@ import pathway as pw
 from pathway.tests.utils import (
     CsvLinesNumberChecker,
     expect_csv_checker,
-    needs_multiprocessing_fork,
     wait_result_with_checker,
 )
 
 
-@needs_multiprocessing_fork
 def test_server(tmp_path: pathlib.Path):
     port = int(os.environ.get("PATHWAY_MONITORING_HTTP_PORT", "20000")) + 10000
     output_path = tmp_path / "output.csv"
@@ -58,7 +56,6 @@ def test_server(tmp_path: pathlib.Path):
     wait_result_with_checker(CsvLinesNumberChecker(output_path, 4), 30)
 
 
-@needs_multiprocessing_fork
 def test_server_customization(tmp_path: pathlib.Path):
     port = int(os.environ.get("PATHWAY_MONITORING_HTTP_PORT", "20000")) + 10001
     output_path = tmp_path / "output.csv"
@@ -99,7 +96,6 @@ def test_server_customization(tmp_path: pathlib.Path):
     wait_result_with_checker(CsvLinesNumberChecker(output_path, 4), 30)
 
 
-@needs_multiprocessing_fork
 def test_server_schema_customization(tmp_path: pathlib.Path):
     port = int(os.environ.get("PATHWAY_MONITORING_HTTP_PORT", "20000")) + 10002
     output_path = tmp_path / "output.csv"
@@ -136,7 +132,6 @@ def test_server_schema_customization(tmp_path: pathlib.Path):
     wait_result_with_checker(CsvLinesNumberChecker(output_path, 4), 30)
 
 
-@needs_multiprocessing_fork
 def test_server_keep_queries(tmp_path: pathlib.Path):
     port = int(os.environ.get("PATHWAY_MONITORING_HTTP_PORT", "20000")) + 10003
     output_path = tmp_path / "output.csv"
@@ -194,17 +189,6 @@ def test_server_fail_on_duplicate_port(tmp_path: pathlib.Path):
         k: int
         v: int
 
-    def target():
-        time.sleep(5)
-        requests.post(
-            f"http://127.0.0.1:{port}/",
-            json={"k": 1, "v": 1},
-        ).raise_for_status()
-        requests.post(
-            f"http://127.0.0.1:{port}/",
-            json={"k": 1, "v": 2},
-        ).raise_for_status()
-
     queries, response_writer = pw.io.http.rest_connector(
         host="127.0.0.1", port=port, schema=InputSchema, delete_completed_queries=False
     )
@@ -227,3 +211,78 @@ def test_server_fail_on_duplicate_port(tmp_path: pathlib.Path):
 
     with pytest.raises(RuntimeError, match="error while attempting to bind on address"):
         pw.run()
+
+
+def test_server_two_endpoints(tmp_path: pathlib.Path):
+    port = int(os.environ.get("PATHWAY_MONITORING_HTTP_PORT", "20000")) + 10004
+    output_path = tmp_path / "output.csv"
+
+    class InputSchema(pw.Schema):
+        query: str
+        user: str
+
+    def uppercase_logic(queries: pw.Table) -> pw.Table:
+        return queries.select(
+            query_id=queries.id, result=pw.apply(lambda x: x.upper(), pw.this.query)
+        )
+
+    def duplicate_logic(queries: pw.Table) -> pw.Table:
+        return queries.select(
+            query_id=queries.id, result=pw.apply(lambda x: x + x, pw.this.query)
+        )
+
+    def target():
+        time.sleep(5)
+        r = requests.post(
+            f"http://127.0.0.1:{port}/duplicate",
+            json={"query": "one", "user": "sergey"},
+        )
+        r.raise_for_status()
+        assert r.text == '"oneone"', r.text
+        r = requests.post(
+            f"http://127.0.0.1:{port}/duplicate",
+            json={"query": "two", "user": "sergey"},
+        )
+        r.raise_for_status()
+        assert r.text == '"twotwo"', r.text
+        r = requests.post(
+            f"http://127.0.0.1:{port}/uppercase",
+            json={"query": "one", "user": "sergey"},
+        )
+        r.raise_for_status()
+        assert r.text == '"ONE"', r.text
+        r = requests.post(
+            f"http://127.0.0.1:{port}/uppercase",
+            json={"query": "two", "user": "sergey"},
+        )
+        r.raise_for_status()
+        assert r.text == '"TWO"', r.text
+
+    webserver = pw.io.http.PathwayWebserver(host="127.0.0.1", port=port)
+
+    uppercase_queries, uppercase_response_writer = pw.io.http.rest_connector(
+        webserver=webserver,
+        schema=InputSchema,
+        route="/uppercase",
+        delete_completed_queries=True,
+    )
+    uppercase_responses = uppercase_logic(uppercase_queries)
+    uppercase_response_writer(uppercase_responses)
+
+    duplicate_queries, duplicate_response_writer = pw.io.http.rest_connector(
+        webserver=webserver,
+        schema=InputSchema,
+        route="/duplicate",
+        delete_completed_queries=True,
+    )
+    duplicate_responses = duplicate_logic(duplicate_queries)
+    duplicate_response_writer(duplicate_responses)
+
+    pw.universes.promise_are_pairwise_disjoint(uppercase_queries, duplicate_queries)
+    all_queries = uppercase_queries.concat(duplicate_queries)
+
+    pw.io.csv.write(all_queries, output_path)
+
+    t = threading.Thread(target=target, daemon=True)
+    t.start()
+    wait_result_with_checker(CsvLinesNumberChecker(output_path, 8), 30)