+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 - `interval_join` can now also work with intervals of zero length.
+- `pw.io.http.rest_connector` now accepts host and port configuration as an instance of the `pw.io.http.PathwayWebserver` class and can now have multiple endpoints running on a single port.
 
 ## [0.7.7] - 2023-12-27